AI_medical_projects/HEALTHWISE-HACKKU25/backend_fastapi/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
motor>=3.3.2
python-dotenv>=1.0.1
pyjwt>=2.8.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
requests>=2.31.0
google-generativeai>=0.5.0
email-validator>=2.1.1
